LayoutTests/http/tests/security/cross-frame-access-protocol.html

<html>
<iframe id="aFrame" rc=""></iframe>
<pre id="console"></pre>
<script>
    function log(s)
    {
        document.getElementById("console").appendChild(document.createTextNode(s + "\n"));
    }

    if (window.layoutTestController) {
        layoutTestController.dumpAsText();
        layoutTestController.waitUntilDone();
    }

    var targetWindow = frames[0];
    if (!targetWindow.document)
        log("FAIL: targetWindow started with no document, we won't know if the test passed or failed.");

    var iframe = document.getElementById('aFrame');
    iframe.src = "https://127.0.0.1:8443/ssl/resources/cross-frame-access-protocol-iframe.html"

    setTimeout(test, 1);

    setTimeout(function() {
        log("FAIL: Cross frame access to https from http was allowed.");
        if (window.layoutTestController)
            layoutTestController.notifyDone();
    }, 5000);

    function test() {
        if (targetWindow.document) {
            setTimeout(test, 1);
            return;
        }

        log("PASS: Cross frame access to https from http was denied!");
        if (window.layoutTestController)
            layoutTestController.notifyDone();
    }
</script>
</html>
