LayoutTests/http/tests/security/cross-frame-access-protocol-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL https://127.0.0.1:8443/ssl/resources/cross-frame-access-protocol-iframe.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-protocol.html. Domains, protocols and ports must match.


PASS: Cross frame access to https from http was denied!
